refactor(CampaignCard): separate Card and Popup components
